+++ src/game/game.ts
@@ -69,9 +69,10 @@ export class Game {
         this.player.addCard();
 
         if (this.player.score > Rules.BLACKJACK) {
-            this.player.displayResult(this.player.bust);
-
-            setTimeout(() => this.newRound(), this.NEW_ROUND_DELAY_MS);
+            setTimeout(() => {
+                this.player.displayResult(this.player.bust);
+                setTimeout(() => this.newRound(), this.NEW_ROUND_DELAY_MS);
+            }, 500);
             return;
         }
 
@@ -107,7 +108,7 @@ export class Game {
         this.ripple(this.standButton);
         this.updateButtons(true, false, [this.hitButton, this.standButton]);
         this.dealer.flipCard();
-        this.startDealersTurn();
+        setTimeout(() => this.startDealersTurn(), 500);
     }
 
     private startDealersTurn(): void {