+++ CHANGELOG.md
@@ -6,7 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [0.5.0] - 10/08/2025
-- Added how to play functionality.
+
+### Added
+- How to play functionality.
+
+### Changed
+- Result message delay increment.
 
 ## [0.4.0] - 09/08/2025