+++ CHANGELOG.md
@@ -5,7 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.1.1] - 15/08/2025
+## [1.2.1] - 15/08/2025
+
+### Added
+- Betting chips.
+
+## [1.1.1] - 14/08/2025
 
 ### Added
 - Double down player action feature.